+++ CHANGELOG.md
@@ -9,6 +9,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- `interface HttpApiConfig`
+- `interface WsApiConfig`
 - `isValidUsernameScalar()`
 - `isValidNameScalar()`
 - `isValidBioScalar()`
@@ -21,7 +23,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
-- `class QueriesApi`, `class MutationsApi`, `class SubscriptionsApi`, and `class RestApi` no longer throw scalar errors such as `class BioScalarError`. You can validate the input manually using the new validation functions such as `isValidBioScalar()` instead. Note that `RestApi.postPicMessage()` still throws a `class MessageTextScalarError`.
+- Make functions previously in `class QueriesApi`, `class MutationsApi`, `class SubscriptionsApi`, and `class RestApi` as top-level functions with a different function signature but the same functionality. These functions no longer throw scalar errors such as `class BioScalarError`. You can validate the input manually using the new validation functions such as `isValidBioScalar()` instead. Note that `postPicMessage()` still throws a `class MessageTextScalarError` because the API server explicitly returns it as one of its responses.
+- Change the function signature of `queryOrMutate()`.
+- Change the function signature of `subscribe()`.
+- `interface GraphQlResponse`
 
 ### Fixed
 
@@ -31,6 +36,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Drop support for Node.js.
 - Drop support for React Native.
+- `class QueriesApi`
+- `class MutationsApi`
+- `class SubscriptionsApi`
+- `class RestApi`
 - `class UuidScalarError`
 - `class PasswordScalarError`
 - `class BioScalarError`
@@ -39,6 +48,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `class UsernameScalarError`
 - `class NameScalarError`
 - `class DateTimeScalarError`
+- `type HttpProtocol`
+- `type WebSocketProtocol`
 
 ## [0.8.0](https://github.com/neelkamath/omni-chat-js/releases/tag/v0.7.0) - 2020-03-20